requirements.txt
asgiref==3.3.1
Django==3.1.5
django-crispy-forms==1.11.0
django-jazzmin==2.4.4
django-leaflet==0.27.1
django-osm-field==0.4.1
Pillow==8.1.0
pytz==2020.5
sqlparse==0.4.1
whitenoise==5.2.0
